requirements.txt
locust
python-daemon